Library

Library

Public interface

ResumableFunctions.ResumableFunctions — Module.

Main module for ResumableFunctions.jl – C# style generators a.k.a. semi-coroutines for Julia

source

ResumableFunctions.@resumable — Macro.

Macro that transforms a function definition in a finite-statemachine:

Defines a new mutable struct that implements the iterator interface and is used to store the internal state.

Makes this new type callable having following characteristics:

implementents the statements from the initial function definition but;

returns at a @yield statement and;

continues after the @yield statement when called again.

Defines a constructor function that respects the calling conventions of the initial function definition and returns an object of the new type.

source

ResumableFunctions.@yield — Macro.

Macro if used in a @resumable function that returns the expr otherwise throws an error.

source

Internals

Base.IteratorSize — Method.

Implements the iteratorsize method of the iterator interface for a subtype of FiniteStateMachineIterator.

source

ResumableFunctions.BoxedUInt8 — Type.

Mutable struct that contains a single UInt8.

source

ResumableFunctions.FiniteStateMachineIterator — Type.

Abstract type used as base type for the type created by the @resumable macro.

source

Base.eltype — Method.

Implements the eltype method of the iterator interface for a subtype of FiniteStateMachineIterator.

source

ResumableFunctions._is_yield — Method.

Function returning whether an expression is a @yield macro

source

ResumableFunctions.get_args — Method.

Function returning the arguments of a function definition

source

ResumableFunctions.get_param_name — Method.

Function returning the name of a where parameter

source

ResumableFunctions.get_slots — Method.

Function returning the slots of a function definition

source

ResumableFunctions.hasreturnvalue — Method.

Function checking the use of a return statement with value

source

ResumableFunctions.make_arg_any — Method.

Function changing the type of a slot arg of a arg = @yield ret or arg = @yield statement to Any.

source

ResumableFunctions.make_args — Method.

Function returning the args for the type construction.

source

ResumableFunctions.remove_catch_exc — Method.

Function removing the exc symbol of a catch exc statement of a list of slots.

source

ResumableFunctions.transform_arg — Method.

Function that replaces a arg = @yield ret statement by

  @yield ret; 
  arg = arg_

where arg_ is the argument of the function containing the expression.

source

ResumableFunctions.transform_exc — Method.

Function that replaces a @yield ret or @yield statement by

  @yield ret
  _arg isa Exception && throw(_arg)

to allow that an Exception can be thrown into a @resumable function.

source

ResumableFunctions.transform_for — Method.

Function that replaces a for loop by a corresponding while loop saving explicitely the iterator and its state.

source

ResumableFunctions.transform_let — Method.

Function that replaces a variable _fsmi.x in an expression by x where x is a variable declared in a let block.

source

ResumableFunctions.transform_slots — Method.

Function that replaces a variable x in an expression by _fsmi.x where x is a known slot.

source

ResumableFunctions.transform_slots_let — Method.

Function that handles let block

source

ResumableFunctions.transform_try — Method.

Function that replaces a try-catch-finally-end expression having a top level @yield statement in the try part

  try
    before_statements...
    @yield ret
    after_statements...
  catch exc
    catch_statements...
  finally
    finally_statements...
  end

with a sequence of try-catch-end expressions:

  try
    before_statements...
  catch
    catch_statements...
    @goto _TRY_n
  end
  @yield ret
  try
    after_statements...
  catch
    catch_statements...
  end
  @label _TRY_n
  finally_statements...

source

ResumableFunctions.transform_yield — Method.

Function that replaces a @yield ret or @yield statement with

  _fsmi._state = n
  return ret
  @label _STATE_n
  _fsmi._state = 0xff

source

ResumableFunctions.transform_yield — Method.

Function that replaces a @yield ret or @yield statement with

  return ret

source